Specify version to ensure building with sbt gives consistent results
When using sbt it's encouraged to specify the precise version of sbt you
want with a `project/build.properties` file. This allows people with
different versions of sbt to build the project with consistent results.

No matter what version of sbt you have installed, the sbt launcher will
read the `build.properties` file and download the appropriate version
if required.

http://www.scala-sbt.org/0.13/docs/Basic-Def.html#Specifying+the+sbt+version